backend/requirements.txt
# Core
Flask==2.3.3
python-dotenv==1.0.0

# Database
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Authentication
Flask-JWT-Extended==4.5.3
bcrypt==4.0.1
PyJWT==2.8.0

# API & HTTP
requests==2.31.0
gunicorn==21.2.0
httpx==0.25.0

# AI Services
openai==0.28.1
chromadb==0.4.15
sentence-transformers==2.2.2
tokenizers==0.15.0
transformers==4.35.2

# Utilities
tenacity==8.2.3
PyPDF2==3.0.1
python-multipart==0.0.6

# CORS
Flask-CORS==4.0.0

# Rate limiting
Flask-Limiter==3.5.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
responses==0.24.0